implement comsosite dataset image render and downloading whole dataset
